+++ Source/General/Project.vb
@@ -49,6 +49,8 @@ Public Class Project
     Public DemuxAudio As DemuxMode = DemuxMode.All
     Public DemuxChapters As Boolean = True
     Public DemuxVideo As Boolean = False
+    Public DolbyVisionProfile As DolbyVisionProfile = DolbyVisionProfile.Profile81
+    Public ExtractHdrmetadata As HdrmetadataMode = HdrmetadataMode.All
     Public ExtractTimestamps As TimestampsMode = TimestampsMode.VfrOnly
     Public ExtractForcedSubSubtitles As Boolean = True
     Public FileExistAudio As FileExistMode
@@ -57,6 +59,7 @@ Public Class Project
     Public ForcedOutputMod As Integer = 8
     Public ForcedOutputModOnlyIfCropped As Boolean = False
     Public HardcodedSubtitle As Boolean
+    Public HdrmetadataFile As String
     Public ImportVUIMetadata As Boolean = True
     Public ITU As Boolean
     Public LastOriginalSourceFile As String